refactor: 프로필 관련 컴포넌트 리팩토링 (#288)
- deviceType (mobile, desktop) 타입 생성
- deviceType에 따른 프로필 컴포넌트 분기 처리
- 불필요한 코드 및 주석 제거
