+++ src/components/feature/mypage/Profile.tsx
@@ -1,6 +1,6 @@
 import { GetUserMe } from '@/types/api-response/user-response'
 
-import { ProfileDesktopUi, ProfileMobileUi } from './index'
+import ProfileUi from './ui/profile-ui/ProfileUi'
 
 type ProfileProps = {
   nickname: GetUserMe['nickname']
@@ -11,10 +11,18 @@ export default function Profile({ nickname, imageUrl }: ProfileProps) {
   return (
     <>
       <div className="block lg:hidden">
-        <ProfileMobileUi nickname={nickname} imageUrl={imageUrl} />
+        <ProfileUi
+          nickname={nickname}
+          imageUrl={imageUrl}
+          deviceType="mobile"
+        />
       </div>
       <div className="hidden lg:block">
-        <ProfileDesktopUi nickname={nickname} imageUrl={imageUrl} />
+        <ProfileUi
+          nickname={nickname}
+          imageUrl={imageUrl}
+          deviceType="desktop"
+        />
       </div>
     </>
   )
