+++ src/components/feature/mypage/ui/profile-ui/index.ts
@@ -1,13 +1,4 @@
-// src/components/feature/mypage/ui/profile-ui/index.ts
-export { DeleteProfileImageUi, EditProfileImageUi } from './ProfileHoverUi'
 export { default as ProfileImageUi } from './ProfileImageUi'
-export { default as ProfileBackgroundUi } from './ProfileBackgroundUi'
-export { default as ProfileBaseUi } from './ProfileBaseUi'
-export { ProfileDesktopGlowUi, ProfileMobileGlowUi } from './ProfileGlowUi'
-export { ProfileDesktopNameUi, ProfileMobileNameUi } from './ProfileNameUi'
-export {
-  ProfileGenresDesktopUi,
-  ProfileGenresMobileUi,
-} from './ProfileGenresUi'
-export { ProfileDesktopUi, ProfileMobileUi } from './ProfileUi'
+export { default as ProfileBackgroundImage } from './ProfileBackgroundImage'
+export { default as ProfileBaseImage } from './ProfileBaseImage'
 export { default as ProfileImageCropDialogUi } from './ProfileImageCropDialogUi'
